+++ src/component/register/registerEducation.tsx
@@ -11,12 +11,21 @@ import { useContext } from "react";
 import RegisterComponentContext from "@/context/RegisterComponentContext";
 
 const RegisterEducation = () => {
-  const { handleNextClick, handleBackClick, register, isValid, handleSubmit } =
-    useContext<any>(RegisterComponentContext);
-
+  const {
+    handleNextClick,
+    handleBackClick,
+    register,
+    isValid,
+    handleSubmit,
+    nicknameCheck,
+  } = useContext<any>(RegisterComponentContext);
   return (
     <>
-      <GoPageBtn onClick={() => handleBackClick("RegisterBasic")}>
+      <GoPageBtn
+        onClick={() => {
+          handleBackClick("RegisterBasic");
+        }}
+      >
         <LeftIc />
       </GoPageBtn>
       <FormContainerDiv>
